Import https://cpan.metacpan.org/authors/id/T/TI/TIMB/perl5.004m4t2.tar.gz

+++ sv.c
@@ -3216,7 +3216,7 @@ thats_really_all_folks:
     SvCUR_set(sv, bp - (STDCHAR*)SvPVX(sv));	/* set length */
     DEBUG_P(PerlIO_printf(Perl_debug_log,
 	"Screamer: done, len=%ld, string=|%.*s|\n",
-	SvCUR(sv),(int)SvCUR(sv),SvPVX(sv)));
+	(long)SvCUR(sv),(int)SvCUR(sv),SvPVX(sv)));
     }
    else
     {
